Mis à jour du Makefile. Suppression du type Byte car le type Byte créer beaucoup de restriction dans le projet contrairement à d'autre type tel que int (le type int remplace le type Byte). Les calculs importants concernant le Quadtree dans le module Calcul sont complets (vérification à effectuer). Le type Pixel contient maintenant un pointeur vers un type Color. Le type Quadtree ne contient plus de pointeur sur un type Color. Un module Hashtable a été ajouté. Le module Hashtable permet de créer un table de hachage permettant de contenir les listes de Noeuds. Cela permet de racourcir les temps de calcul (évite de parcourir tout un arbre pour chercher un seul élément, notamment les feuilles).

+++ src/Compression.c
@@ -55,7 +55,7 @@ void write_B_W(BitFile* out, Node* node){
 		write_BitFile(out, 1);
 }
 
-void write_unite(BitFile* out, Byte unite){
+void write_unite(BitFile* out, int unite){
 	int i;
 	int bit;
 	int mask;
